+++ docs/changeLog.md
@@ -6,6 +6,15 @@ title: 更新日志
 
 遵循 [Semantic Versioning 2.0.0](https://semver.org/lang/zh-CN/) 语义化版本规范。
 
+### 🚀 2.8.0
+
+`2021-04-19`
+
+- tools.momentToString：优化了对 Invalid date 的处理
+- 新增组件 [DateEasily](/components/date-easily)
+- TdUpload
+  - 新增属性 `showDownLoad`，控制组件预览时，是否有下载按钮。
+
 ### 🚀 2.7.0
 
 `2021-04-13`